requirements.txt
streamlit
google-cloud-aiplatform
magika
gitpython